+++ src/main.js
@@ -1,5 +1,10 @@
 import { createApp } from 'vue'
+import router from './routes/routes'
 import './style.css'
 import App from './App.vue'
+import Header from "./components/Header.vue"
 
-createApp(App).mount('#app')
+const app = createApp(App)
+
+app.use(router)
+app.mount('#app')
